+++ sdk/core/azure-core/src/main/java/com/azure/core/util/Context.java
@@ -57,6 +57,8 @@ public class Context {
     private final Object value;
     private final int contextCount;
 
+    private Map<Object, Object> valuesMap;
+
     /**
      * Constructs a new {@link Context} object.
      *
@@ -215,11 +217,20 @@ public class Context {
         if (key == null) {
             throw LOGGER.logExceptionAsError(new IllegalArgumentException("key cannot be null"));
         }
+
         for (Context c = this; c != null; c = c.parent) {
             if (key.equals(c.key)) {
                 return Optional.ofNullable(c.value);
             }
+
+            // If the contextCount is 1 that means the next parent Context is the NONE Context.
+            // Return Optional.empty now to prevent a meaningless check.
+            if (c.contextCount == 1) {
+                return Optional.empty();
+            }
         }
+
+        // This should never be reached but is required by the compiler.
         return Optional.empty();
     }
 
@@ -256,15 +267,31 @@ public class Context {
      * @return A map containing all values of the context linked-list.
      */
     public Map<Object, Object> getValues() {
-        return getValuesHelper(new HashMap<>());
-    }
-
-    private Map<Object, Object> getValuesHelper(Map<Object, Object> values) {
-        if (key != null) {
-            values.putIfAbsent(key, value);
+        if (valuesMap != null) {
+            return valuesMap;
         }
 
-        return (parent == null) ? values : parent.getValuesHelper(values);
+        if (contextCount == 1) {
+            this.valuesMap = Collections.singletonMap(key, value);
+            return this.valuesMap;
+        }
+
+        Map<Object, Object> map = new HashMap<>((int) Math.ceil(contextCount / 0.75F));
+
+        for (Context pointer = this; pointer != null; pointer = pointer.parent) {
+            if (pointer.key != null) {
+                map.putIfAbsent(pointer.key, pointer.value);
+            }
+
+            // If the contextCount is 1 that means the next parent Context is the NONE Context.
+            // Break out of the loop to prevent a meaningless check.
+            if (pointer.contextCount == 1) {
+                break;
+            }
+        }
+
+        this.valuesMap = Collections.unmodifiableMap(map);
+        return this.valuesMap;
     }
 
     /**
@@ -275,11 +302,16 @@ public class Context {
     Context[] getContextChain() {
         Context[] chain = new Context[contextCount];
 
-        Context pointer = this;
         int chainPosition = contextCount - 1;
-        while (pointer != null && chainPosition >= 0) {
+
+        for (Context pointer = this; pointer != null; pointer = pointer.parent) {
             chain[chainPosition--] = pointer;
-            pointer = pointer.parent;
+
+            // If the contextCount is 1 that means the next parent Context is the NONE Context.
+            // Break out of the loop to prevent a meaningless check.
+            if (pointer.contextCount == 1) {
+                break;
+            }
         }
 
         return chain;