+++ sdk/recoveryservices/azure-resourcemanager-recoveryservices/src/main/java/com/azure/resourcemanager/recoveryservices/models/VaultCertificateResponse.java
@@ -58,7 +58,7 @@ public interface VaultCertificateResponse {
             /**
              * Specifies resourceGroupName, vaultName.
              *
-             * @param resourceGroupName The name of the resource group where the recovery services vault is present.
+             * @param resourceGroupName The name of the resource group. The name is case insensitive.
              * @param vaultName The name of the recovery services vault.
              * @return the next definition stage.
              */