+++ sdk/spring/spring-cloud-azure-actuator-autoconfigure/src/main/java/com/azure/spring/cloud/actuator/autoconfigure/implementation/cosmos/CosmosHealthConfiguration.java
@@ -4,7 +4,7 @@
 package com.azure.spring.cloud.actuator.autoconfigure.implementation.cosmos;
 
 import com.azure.cosmos.CosmosAsyncClient;
-import com.azure.spring.cloud.actuator.cosmos.CosmosHealthIndicator;
+import com.azure.spring.cloud.actuator.implementation.cosmos.CosmosHealthIndicator;
 import com.azure.spring.cloud.autoconfigure.implementation.cosmos.AzureCosmosAutoConfiguration;
 import com.azure.spring.cloud.autoconfigure.implementation.cosmos.properties.AzureCosmosProperties;
 import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;