+++ sdk/spring/spring-cloud-azure-actuator-autoconfigure/src/main/java/com/azure/spring/cloud/actuator/autoconfigure/implementation/storage/StorageBlobHealthConfiguration.java
@@ -3,7 +3,7 @@
 
 package com.azure.spring.cloud.actuator.autoconfigure.implementation.storage;
 
-import com.azure.spring.cloud.actuator.storage.StorageBlobHealthIndicator;
+import com.azure.spring.cloud.actuator.implementation.storage.StorageBlobHealthIndicator;
 import com.azure.spring.cloud.autoconfigure.implementation.storage.blob.AzureStorageBlobAutoConfiguration;
 import com.azure.storage.blob.BlobServiceAsyncClient;
 import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;